dns: fix non-primary coredns failures
Due to an oversight (the use of a hostname rather than an IP address), all DNS instances for non-primary nodes were crash-looping.

This was not initially noticed due to the fallback behavior being to forward to the cluster primary. The only impact from this issue was slower DNS query times on non-primary nodes due to the need for retries - an annoying degradation, but not an outage.