dns: cleanup coredns state and install process
This change makes the DNS state more robust to changes to various files to ensure that CoreDNS will be properly restarted and also fixes an annoying issue with the SLS file managing /usr/local/bin, which is a situation to be avoided.

This change was deployed in a piecemeal fashion due to the nature of the change to the CoreDNS install location, with several intermediate actions not included here.